Avoid naming conflict with core atomspace
AtomSpace now implements a mre generic, more powerful ImpulseLink
desiged to make AS-MOSES more generic, and I think it will also
make it run a lot faser, because it can do entire table columns
in one go.

However, there is now a naming cnflict, so rename this one, for now.
I plan to eventually remove this one and use the other instead.